% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fun.R
\name{cv_fit}
\alias{cv_fit}
\title{Cross-Validation Fit for WA Regression (Internal)}
\usage{
cv_fit(
  model,
  data,
  cluster = NULL,
  id,
  wr,
  wd,
  link,
  cens_mod,
  degree,
  time_range,
  L,
  K = 10,
  t_span
)
}
\arguments{
\item{model}{A survival model formula (e.g. \code{Surv(time, type) ~ Z1 + Z2}).}

\item{data}{A data frame containing the data used for fitting the model.}

\item{cluster}{Optional. A string specifying the column name for cluster ID (if applicable).}

\item{id}{A string specifying the column name for subject ID.}

\item{wr}{A numeric vector of weights for recurrent events.}

\item{wd}{A numeric value representing the weight for the terminal event.}

\item{link}{Link function for while alive loss rate.}

\item{cens_mod}{A survival model formula for the censoring distribution (e.g. \code{Surv(time, delta) ~ Z2}).}

\item{degree}{A non-negative integer specifying the polynomial degree for the spline.}

\item{time_range}{A numeric vector specifying the spline knot locations.}

\item{L}{A number of knots for cross-validation.}

\item{K}{An integer indicating the number of folds for cross-validation (default is 10).}

\item{t_span}{A numeric vector of time points at which the prediction error is evaluated.}
}
\value{
A numeric value representing the average prediction error across folds.
}
\description{
This internal function performs K-fold cross validation to assess the prediction error
of the WA regression model. For each fold, the model is fitted on the training data using
the specified spline parameters and then the prediction error is computed on the test data.
The final cross-validation error is the average prediction error across all folds.
}
\keyword{internal}
