% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fun.R
\name{est_eq}
\alias{est_eq}
\title{Compute the Estimating Equation (Internal)}
\usage{
est_eq(beta, fd_dt, cov_vars, link)
}
\arguments{
\item{beta}{A numeric vector of coefficients.}

\item{fd_dt}{A data frame containing the required columns (e.g., \code{V_tau}, \code{G_X_tau},
\code{L}, \code{X_tau}) along with the expanded covariate columns.}

\item{cov_vars}{A character vector of covariate names used to construct the spline-expanded
covariate columns.}

\item{link}{link function for the while-alive hazard rate, default is "log"}
}
\value{
A numeric vector representing the estimating equation \eqn{U}.
}
\description{
This internal function computes the estimating equation \eqn{U} based on the expanded covariate
matrix from \code{fd_dt} and the coefficient vector \code{beta}. It extracts the relevant
spline-expanded covariate columns (using \code{cov_vars}) from \code{fd_dt}, computes the linear
predictor, and then calculates \eqn{U} as the column means of the row-wise product of the covariate
matrix and a scalar term.
}
\keyword{internal}
