% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WA_fit.R
\name{summary.WA_obj}
\alias{summary.WA_obj}
\title{Summary Method for WA_obj Objects}
\usage{
\method{summary}{WA_obj}(object)

\method{summary}{WA_obj}(object)
}
\arguments{
\item{object}{A WA_obj object returned by WA_reg.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Invisibly returns the summary table.

Invisibly returns the summary table.
}
\description{
This function provides a summary of a WA_obj object including the spline knots,
the primary model, the censoring model, the degree of the spline, and a summary
table of the estimated parameters (along with their standard errors, z-values,
and two-sided p-values).

This function provides a summary of a WA_obj object including the spline knots,
the primary model, the censoring model, the degree of the spline, and a summary
table of the estimated parameters (along with their standard errors, z-values,
and two-sided p-values).
}
